refactor: 함수 단위에 맞게 리팩토링
Dto는 View에서 Domain에 대한 의존을 끊기 위해서 존재하는 객체이다.
따라서 Domain에서 Dto를 의존하면 안된다.

Controller에서 흐름을 잘보이게 하기 위해서 inputView에 입력요청 -> 데이터 처리 -> outputView에 출력요청 순서로 나열
각 의미별로 개행을 통해 분리하기 쉽도록 리팩토링
